+++ gthumb/gth-test-simple.h
@@ -40,10 +40,10 @@ typedef struct _GthTestSimple         GthTestSimple;
 typedef struct _GthTestSimplePrivate  GthTestSimplePrivate;
 typedef struct _GthTestSimpleClass    GthTestSimpleClass;
 
-typedef gint64 (*GthTestGetData) (GthTest        *test,
-				  GthFileData    *file,
-				  gpointer       *data,
-				  GDestroyNotify *data_destroy_func);
+typedef int (*GthTestGetData) (GthTest        *test,
+			       GthFileData    *file,
+			       gpointer       *data,
+			       GDestroyNotify *data_destroy_func);
 
 typedef enum {
 	GTH_TEST_DATA_TYPE_NONE,
